requirements.txt
fastapi
uvicorn
pytest
httpx
psycopg2-binary
python-dotenv
pydantic-settings
SQLAlchemy